# docs/swagger.yml
openapi: 3.0.3
info:
  title: Financial Assistance Application API
  version: 1.0.0
  description: This API enables the management and processing of financial assistance applications.
    It allows students to apply for grants, and provides system admins and other roles access to manage and track the status of applications.
    The API supports roles such as students, financial assistance coordinators, and various offices responsible for managing the budget, processing payments, and approving grants.
    Key features include application submission, status updates, and administrative control over grant approvals and disbursements.
servers:
  - url: 'https://{environment}/api/v1'
    description: Server url
...
security:
  - BearerAuth: []
paths:
...
  /sponsorships/coordinator:
    get:
      tags:
        - Coordinator Sponsorship
      summary: Get all sponsorship
      responses:
        '200':
          description: A list of sponsorship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllSponsorshipSuccessResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
    post:
      tags:
        - Coordinator Sponsorship
      summary: Register a new Sponsorship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SponsorshipPayload'
      responses:
        '201':
          description: Resource successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSponsorshipSuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/coordinator/{sponsorshipId}:
    get:
      tags:
        - Coordinator Sponsorship
      parameters:
        - name: sponsorshipId
          in: path
          required: true
          description: The ID of the Sponsorship
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      summary: Get one Sponsorship
      responses:
        '200':
          description: A details of Sponsorship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSponsorshipSuccessResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
    put:
      tags:
        - Coordinator Sponsorship
      parameters:
        - name: sponsorshipId
          in: path
          required: true
          description: The ID of the Sponsorship
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      summary: Register a new Sponsorship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SponsorshipPayload'
      responses:
        '201':
          description: Resource successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSponsorshipSuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
    delete:
      tags:
        - Coordinator Sponsorship
      parameters:
        - name: sponsorshipId
          in: path
          required: true
          description: The ID of the Sponsorship
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      summary: Delete a Sponsorship
      responses:
        '201':
          description: Resource successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessDeleteResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/coordinator/students/{studentId}/status:
    put:
      tags:
        - Coordinator Sponsorship
      parameters:
        - name: studentId
          in: path
          required: true
          description: The ID of the student
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      summary: Register a new Sponsorship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SponsorshipChangeStatusPayload'
      responses:
        '201':
          description: Resource successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSponsorshipChangeStatus'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/student/available:
    get:
      tags:
        - Student Sponsorship
      summary: Get all sponsorship available for the student
      responses:
        '200':
          description: A list of sponsorship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllSponsorshipSuccessResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
...
  /sponsorships/student/:
    post:
      tags:
        - Student Sponsorship
      summary: Register student to a Sponsorship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplySponsorshipPayload'
      responses:
        '201':
          description: Resource successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetApplySponsorshipSuccessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
...
  /sponsorships/update-criterion/{sponsorshipId}:
    put:
      tags:
        - Coordinator Sponsorship
      parameters:
        - name: sponsorshipId
          in: path
          required: true
          description: The ID of the sponsorship
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriterionPayload'
      responses:
        '201':
          description: Resource successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCriterionSuccessResponse'
  /sponsorships/criterion-category:
    get:
      tags:
        - Coordinator Sponsorship
      summary: Get all criterion category
      responses:
        '200':
          description: A list of criterion category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllCriteriaCategorySuccessResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/criterion-category/data-sources:
    get:
      tags:
        - Coordinator Sponsorship
      summary: Get all criterion category
      responses:
        '200':
          description: A list of criterion category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDataSourceSuccessResponse'
        '403':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/bulk-upsert-custom-input:
    post:
      tags:
        - Coordinator Sponsorship
      requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkCustomInputPayload'
      responses:
        '201':
          description: Resource successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkCustomInputSuccessResponse'
...
  /sponsorships/custom-input:
    get:
      tags:
        - Coordinator Sponsorship
      parameters:
        - name: studentId
          in: query
          required: false
          description: The ID of the student
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
        - name: sponsorshipId
          in: query
          required: false
          description: The ID of the sponsorship
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
        - name: sponsorshipCriterionId
          in: query
          required: false
          description: The ID of the sponsorship criterion
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      responses:
        '201':
          description: Resource successfully get
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBulkCustomInputSuccessResponse'
...
  /sponsorships/student/available/{studentId}:
    get:
      tags:
        - Student Sponsorship
      summary: Get available sponsorships for a specific student
      security:
        - BearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          description: The ID of the student
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      responses:
        '200':
          description: A list of available sponsorships for the student
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllSponsorshipSuccessResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/student/{sponsorshipId}:
    get:
      tags:
        - Student Sponsorship
      summary: Get one student sponsorship application
      security:
        - BearerAuth: []
      parameters:
        - name: sponsorshipId
          in: path
          required: true
          description: The ID of the sponsorship application
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      responses:
        '200':
          description: Details of a student sponsorship application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetApplySponsorshipSuccessResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/student/my-sponsorships/{studentId}:
    get:
      tags:
        - Student Sponsorship
      summary: Get all sponsorships for a student
      security:
        - BearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          description: The ID of the student
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      responses:
        '200':
          description: A list of the student's sponsorship applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllApplySponsorshipSuccessResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/student/{studentId}/cancel/{sponsorshipId}:
    delete:
      tags:
        - Student Sponsorship
      summary: Cancel a sponsorship application (only allowed at POOLING stage)
      security:
        - BearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          description: The ID of the student
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
        - name: sponsorshipId
          in: path
          required: true
          description: The ID of the sponsorship
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      responses:
        '200':
          description: Sponsorship application cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessDeleteResponse'
        '400':
          description: Application not found or not in POOLING stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/applicants/by-stage:
    get:
      tags:
        - Coordinator Sponsorship
      summary: Get all applicants filtered by application stage
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: applicationStage
          schema:
            type: string
          description: Filter by application stage
        - in: query
          name: applicationStatus
          schema:
            type: string
          description: Filter by application status
        - in: query
          name: sponsorshipId
          schema:
            type: string
          description: Filter by sponsorship ID
        - in: query
          name: search
          schema:
            type: string
          description: Search term for filtering applicants
        - in: query
          name: sort
          schema:
            type: string
          description: Field to sort by
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
          description: Number of records to skip
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
          description: Maximum number of records to return
      responses:
        '200':
          description: A list of applicants filtered by stage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllApplySponsorshipSuccessResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
  /sponsorships/rank-student/{sponsorshipId}:
    get:
      tags:
        - Coordinator Sponsorship
      summary: Rank students for a sponsorship using SAW score results
      security:
        - BearerAuth: []
      parameters:
        - name: sponsorshipId
          in: path
          required: true
          description: The ID of the sponsorship
          schema:
            type: string
            example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
      responses:
        '200':
          description: Ranked students with SAW scores
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Resource Successfully Retrieved!
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        studentId:
                          type: string
                          example: fa8d0ba7-3bd8-4a30-8127-82ca571d1bd0
                        studentName:
                          type: string
                          example: John Doe
                        sawScore:
                          type: number
                          example: 0.85
                        rank:
                          type: integer
                          example: 1
                  generatedAt:
                    type: integer
                    example: 1738335123028
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalErrorResponse'
...
components:
  schemas:
...
    SponsorshipPayload:
      type: object
      properties:
        name:
          type: string
          example: Sponsor number 1
        sponsorId:
          type: string
          example: a2815764-f067-11ef-ae8c-f4b520508eb0
        academicYearId:
          type: string
          example: a2815764-f067-11ef-ae8c-f4b520508eb0
        durationFrom:
          type: string
          example: "2023-08-02 00:00:00"
        durationTo:
          type: string
          example: "2023-09-02 00:00:00"
        batchNumber:
          type: integer
          example: 1
        limit:
          type: integer
          example: 1
        slot:
          type: integer
          example: 1
        fundAllocation:
          type: number
          example: 1000
        status:
          type: integer
          example: 1
        sponsorshipSchool:
          type: array
          items:
            type: string
            example: a2815764-f067-11ef-ae8c-f4b520508eb0
        sponsorshipRequirements:
          type: array
          items:
            type: string
            example: a2815764-f067-11ef-ae8c-f4b520508eb0
...
    InternalErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          example: INTERNAL-SERVER-ERror
        generatedAt:
          type: integer
          example: 1738335047588
        description:
          type: string
          example: Server Errror Occured
        errorDetails:
          type: string
          example: Server Errror Occured
    ForbiddenErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          example: FORBIDDEN
        generatedAt:
          type: integer
          example: 1738335047588
        description:
          type: string
          example: The resource requested is forbidden for your API Key.
        errorDetails:
          type: string
          example: Invalid API Token
...
    GetAllSponsorshipSuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Request Successfully Accepted!
        data:
          type: array
          items: 
            $ref: '#/components/schemas/SponsorshipResponse' 
...
    GetSponsorshipSuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Request Successfully Accepted!
        data:
          $ref: '#/components/schemas/SponsorshipResponse' 
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT